net/proxy use perfect forwarding for templated callbacks
This workarounds of a caveat in the boost asio library, where async operations prepare buffers and move self in a single line, see read_op [1] for example. When the callback is a value type instead of a reference, self could be moved before the buffer is accessed, resulting in a use-after-move error.

[1] https://github.com/boostorg/asio/blob/d6e7b5a547daaddfd19c548d2f602cb5b15361df/include/boost/asio/impl/read.hpp#L398